refactor(cli): share body parameter naming helpers
