+++ internal/pipeline/toolsmanifest.go
@@ -9,11 +9,11 @@ import (
 	"path/filepath"
 	"sort"
 	"strings"
-	"unicode"
 
 	"github.com/mvanhorn/cli-printing-press/v4/internal/mcpdesc"
 	"github.com/mvanhorn/cli-printing-press/v4/internal/mcpoverrides"
 	"github.com/mvanhorn/cli-printing-press/v4/internal/naming"
+	"github.com/mvanhorn/cli-printing-press/v4/internal/paramnames"
 	"github.com/mvanhorn/cli-printing-press/v4/internal/spec"
 )
 
@@ -447,7 +447,7 @@ func manifestBodyParams(ep spec.Endpoint) []manifestBodyParam {
 		}
 		return params
 	}
-	body := manifestFlattenCollidingBodyFields(ep.Body)
+	body := paramnames.FlattenCollidingBodyFields(ep.Body)
 	params := make([]manifestBodyParam, 0, len(body))
 	collectManifestBodyParams(&params, body, 0, "", nil)
 	return params
@@ -465,11 +465,11 @@ func collectManifestBodyParams(params *[]manifestBodyParam, body []spec.Param, d
 				continue
 			}
 			nextPath := append(append([]string(nil), bodyPath...), p.BodyWireName())
-			collectManifestBodyParams(params, p.Fields, depth+1, manifestJoinFlag(flagPrefix, manifestPublicFlagName(p)), nextPath)
+			collectManifestBodyParams(params, p.Fields, depth+1, naming.JoinFlag(flagPrefix, paramnames.PublicFlagName(p)), nextPath)
 			continue
 		}
 		if flagPrefix != "" {
-			p.FlagName = manifestJoinFlag(flagPrefix, manifestPublicFlagName(p))
+			p.FlagName = naming.JoinFlag(flagPrefix, paramnames.PublicFlagName(p))
 			p.Aliases = nil
 		}
 		wireName := p.BodyWireName()
@@ -478,88 +478,6 @@ func collectManifestBodyParams(params *[]manifestBodyParam, body []spec.Param, d
 		}
 		*params = append(*params, manifestBodyParam{Param: p, WireName: wireName})
 	}
-}
-
-func manifestFlattenCollidingBodyFields(body []spec.Param) []spec.Param {
-	counts := manifestCountBodyLeaves(body, "")
-	for _, n := range counts {
-		if n > 1 {
-			return manifestClearCollidingParents(body, "", counts)
-		}
-	}
-	return body
-}
-
-func manifestCountBodyLeaves(params []spec.Param, prefix string) map[string]int {
-	counts := map[string]int{}
-	var walk func([]spec.Param, string)
-	walk = func(ps []spec.Param, pfx string) {
-		for _, p := range ps {
-			ident := pfx + manifestToCamel(manifestParamIdent(p))
-			if p.Type == "object" && len(p.Fields) > 0 {
-				walk(p.Fields, ident)
-				continue
-			}
-			counts[ident]++
-		}
-	}
-	walk(params, prefix)
-	return counts
-}
-
-func manifestClearCollidingParents(params []spec.Param, prefix string, counts map[string]int) []spec.Param {
-	out := make([]spec.Param, len(params))
-	copy(out, params)
-	for i := range out {
-		p := &out[i]
-		if p.Type != "object" || len(p.Fields) == 0 {
-			continue
-		}
-		ident := prefix + manifestToCamel(manifestParamIdent(*p))
-		if manifestSubtreeHasCollidingLeaf(p.Fields, ident, counts) {
-			p.Fields = nil
-			continue
-		}
-		p.Fields = manifestClearCollidingParents(p.Fields, ident, counts)
-	}
-	return out
-}
-
-func manifestSubtreeHasCollidingLeaf(params []spec.Param, prefix string, counts map[string]int) bool {
-	for _, p := range params {
-		ident := prefix + manifestToCamel(manifestParamIdent(p))
-		if p.Type == "object" && len(p.Fields) > 0 {
-			if manifestSubtreeHasCollidingLeaf(p.Fields, ident, counts) {
-				return true
-			}
-			continue
-		}
-		if counts[ident] > 1 {
-			return true
-		}
-	}
-	return false
-}
-
-func manifestParamIdent(p spec.Param) string {
-	if p.IdentName != "" {
-		return p.IdentName
-	}
-	return p.Name
-}
-
-func manifestPublicFlagName(p spec.Param) string {
-	if p.FlagName != "" {
-		return p.FlagName
-	}
-	return manifestFlagName(manifestParamIdent(p))
-}
-
-func manifestJoinFlag(prefix, name string) string {
-	if prefix == "" {
-		return name
-	}
-	return prefix + "-" + name
 }
 
 func uniqueManifestParamName(name string, used map[string]struct{}) string {
@@ -577,51 +495,6 @@ func uniqueManifestParamName(name string, used map[string]struct{}) string {
 			return candidate
 		}
 	}
-}
-
-func manifestToCamel(s string) string {
-	s = strings.TrimLeft(s, "$")
-	parts := strings.FieldsFunc(s, func(r rune) bool {
-		return !unicode.IsLetter(r) && !unicode.IsDigit(r)
-	})
-	for i, p := range parts {
-		if len(p) > 0 {
-			parts[i] = strings.ToUpper(p[:1]) + p[1:]
-		}
-	}
-	result := strings.Join(parts, "")
-	if len(result) > 0 && !unicode.IsLetter(rune(result[0])) {
-		result = "V" + result
-	}
-	return result
-}
-
-func manifestFlagName(name string) string {
-	name = strings.TrimLeft(name, "$")
-	var b strings.Builder
-	runes := []rune(name)
-	for i, r := range runes {
-		if !unicode.IsLetter(r) && !unicode.IsDigit(r) {
-			if b.Len() > 0 {
-				b.WriteByte('-')
-			}
-			continue
-		}
-		if i > 0 && unicode.IsUpper(r) {
-			prev := runes[i-1]
-			if unicode.IsLower(prev) || unicode.IsDigit(prev) {
-				b.WriteByte('-')
-			} else if unicode.IsUpper(prev) && i+1 < len(runes) && unicode.IsLower(runes[i+1]) {
-				b.WriteByte('-')
-			}
-		}
-		b.WriteRune(unicode.ToLower(r))
-	}
-	result := b.String()
-	for strings.Contains(result, "--") {
-		result = strings.ReplaceAll(result, "--", "-")
-	}
-	return strings.Trim(result, "-")
 }
 
 // reservedManifestParamNames seeds generator-reserved public names only.
